Make epm check and exit if patch isn't properly applied
Use getProgramInvocationOutput instead of runProgramInvocation as it
will exit properly if the exit status is not ExitSuccess. A sample
contirved demo of it's behavior:

```
epm: Error: some packages failed to install:
old-time-1.1.0.3 failed while unpacking the package. The exception was:
user error ('/usr/bin/git' exited with an error:
git: 'applydfkja' is not a git command. See 'git --help'.)
```

Fixes #4.